+++ rosjava_utils/rosjava_utils/src/main/java/gov/dot/fhwa/saxton/carma/rosjava_utils/MobilityHelper.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package gov.dot.fhwa.saxton.carma.rosutils;
+package gov.dot.fhwa.saxton.carma.rosjava_utils;
 
 import java.util.ArrayList;
 import java.util.List;
