+++ rosjava_utils/rosjava_utils/src/main/java/gov/dot/fhwa/saxton/carma/rosjava_utils/RosServiceResponse.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package gov.dot.fhwa.saxton.carma.rosutils;
+package gov.dot.fhwa.saxton.carma.rosjava_utils;
 
 import org.ros.node.service.ServiceResponseListener;
 import org.ros.exception.RemoteException;
