+++ rosjava_utils/rosjava_utils/src/test/java/gov/dot/fhwa/saxton/carma/rosutils/RosTest.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package gov.dot.fhwa.saxton.carma.rosutils;
+package gov.dot.fhwa.saxton.carma.rosjava_utils;
 
 import static org.junit.Assert.assertTrue;
